// src/types/Anilist/AnimeStreamingLlinks.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface StreamingLinks {
    headers:  Headers;
    sources:  Source[];
    download: string;
}

export interface Headers {
    Referer: string;
}

export interface Source {
    url:     string;
    isM3U8:  boolean;
    quality: string;
}